+++ src/main/java/org/d2ab/iterator/SteppingIterator.java
@@ -20,11 +20,11 @@ import java.util.NoSuchElementException;
 
 public class SteppingIterator<T> implements Iterator<T> {
 	private final Iterator<T> iterator;
-	private final int step;
+	private final long step;
 	private boolean gotNext;
 	private T next;
 
-	public SteppingIterator(Iterator<T> iterator, int step) {
+	public SteppingIterator(Iterator<T> iterator, long step) {
 		this.iterator = iterator;
 		this.step = step;
 	}
@@ -40,7 +40,7 @@ public class SteppingIterator<T> implements Iterator<T> {
 		next = iterator.next();
 
 		// skip steps
-		int i = step;
+		long i = step;
 		while (--i > 0 && iterator.hasNext())
 			iterator.next();
 		gotNext = true;
